Provide incremental readdir
Since current implementation iterates over whole odb through a
callback, separate thread is used for reading.

Two threads communicate through a pipe. In the future it will
probbably be switch to a blocking queue, but for now, this is
a simplistic approach.

Offset in readdir is used only to determine whether dir handle is
in a valid state. It's not possible to load an arbitrary offset.

Regarding thread safety, I couldn't find whether FUSE will ensure
that file handles are called exclusively or not. For now I assume
that it does.